feat: display visitor date/time in real time
