+++ src/App.jsx
@@ -1,6 +1,7 @@
 import Navigation from './components/Navigation';
 import NookCard from './components/NookCard';
 import DailyBday from './components/DailyBday';
+import VisitorCard from './components/VisitorCard';
 
 import { useState, useEffect } from 'react'
 import { getEvents } from './api.js'
@@ -11,7 +12,6 @@ import nookling from "./assets/nookling.png";
 import player from "./assets/player.png";
 
 import './App.css'
-import VisitorCard from './components/VisitorCard';
 
 function App() {
